requirements.txt
# Core SMA Framework
mesa==2.2.4

# Data Processing
numpy==1.26.4
pandas==2.2.1

# Routing Algorithms
networkx==3.2.1

# Machine Learning (Q-Learning)
scikit-learn==1.4.1.post1

# Visualization
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# Database
psycopg[binary]==3.2.4
psycopg-pool==3.2.4

# SUMO (visualisation temps réel)
eclipse-sumo>=1.20.0
traci>=1.20.0
sumolib>=1.20.0
tabulate==0.9.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1

# Logging
loguru==0.7.2

# Testing
pytest==8.1.1
pytest-cov==4.1.0

# Development
black==24.3.0
flake8==7.0.0
